+++ public/command-palette/index.html
@@ -1,0 +1,78 @@
+<body>
+  <link rel="stylesheet" href="/command-palette/style.css">
+  <script type="module" src="/command-palette/index.js"></script>
+  
+  <sl-dialog class="command-palette" label="Command Palette" style="--width: 600px">
+    <form class="palette-form">
+      <sl-input class="palette-filter" autofocus placeholder="Search"></sl-input>
+      <sl-menu>
+        <sl-menu-item value="set-title">Set Title</sl-menu-item>
+        <sl-menu-item value="save">Save Logs To Disk</sl-menu-item>
+        <sl-menu-item value="help">Help</sl-menu-item>
+      </sl-menu>
+    </form>
+  
+    <form class="hide title-form">
+      <sl-input label="New Title" autofocus placeholder="New page title"></sl-input>
+      <div class="form-footer">
+        <sl-button type="submit">Change Title</sl-button>
+      </div>
+    </form>
+    <form class="hide help-menu">
+      <dl>
+        <dt>My logs aren't showing!</dt>
+        <dd>
+          <p>Your logs may be outputting to stderr. Try using bash redirection.</p>
+          <p>
+            <code>my-program <strong>2>&1</strong> | logpipe</code>
+          </p>
+        </dd>
+  
+        <dt>How do I specify a port?</dt>
+        <dd>
+          <p>The command line supports the options <code>--port</code> and <code>--title</code> for specifying these
+            things.</p>
+          <p>
+            <code>my-program | logpipe --port 8080 --title "Backend System"</code>
+          </p>
+        </dd>
+  
+        <dt>How do I copy a log or timestamp?</dt>
+        <dd>
+          <p>Try out the right-click menu! Right click a log for a list of options.</p>
+        </dd>
+  
+        <dt>How do I jump to a log's source location?</dt>
+        <dd>
+          <p>Also in the right-click menu!</p>
+        </dd>
+  
+        <dt>How do I search for all URLs?</dt>
+        <dd>
+          <p>Use the logpipe query language!</p>
+          <p>
+            <code>@@url</code>
+          </p>
+        </dd>
+  
+        <dt>What are the supported query types?</dt>
+        <dd>
+          <p>Here is a query which includes all of them!</p>
+          <p>
+            <code class="multiline">@@string,number,tag,url,ip,error,file,http-method,key,value,keyword</code>
+          </p>
+        </dd>
+  
+        <dt>What are the keyboard shortcuts?</dt>
+        <dd>
+          <p><kbd>Cmd</kbd>+<kbd>k</kbd> - Open the command palette</p>
+          <p><kbd>/</kbd> - Focus filter box</p>
+        </dd>
+      </dl>
+  
+      <div class="form-footer">
+        <sl-button type="submit">Got It!</sl-button>
+      </div>
+    </form>
+  </sl-dialog>
+</body>
