+++ public/command-palette/style.css
@@ -1,0 +1,28 @@
+sl-dialog form:not(.palette-form, .hide) {
+  display: grid;
+  gap: var(--sl-spacing-medium);
+}
+
+form.help-menu {
+  dl {
+    padding: 0 var(--sl-spacing-x-large);
+  }
+  dt {
+    color: var(--sl-color-neutral-900);
+  }
+  dd {
+    color: var(--sl-color-neutral-700);
+  }
+  code {
+    background: var(--sl-color-neutral-300);
+    padding: var(--sl-spacing-2x-small) var(--sl-spacing-x-small);
+    border-radius: var(--sl-border-radius-small);
+  }
+  code.multiline {
+    display: block;
+  }
+}
+.form-footer {
+  display: flex;
+  justify-content: flex-end;
+}
