+++ public/style.css
@@ -231,45 +231,6 @@ header {
   }
 }
 
-.contextmenu {
-  position: absolute;
-  left: calc(var(--x) * 1px);
-  top: calc(var(--y) * 1px);
-  display: none;
-  &.show {
-    display: block;
-  }
-}
-
-sl-dialog form:not(.palette-form, .hide) {
-  display: grid;
-  gap: var(--sl-spacing-medium);
-}
-
-form.help-menu {
-  dl {
-    padding: 0 var(--sl-spacing-x-large);
-  }
-  dt {
-    color: var(--sl-color-neutral-900);
-  }
-  dd {
-    color: var(--sl-color-neutral-700);
-  }
-  code {
-    background: var(--sl-color-neutral-300);
-    padding: var(--sl-spacing-2x-small) var(--sl-spacing-x-small);
-    border-radius: var(--sl-border-radius-small);
-  }
-  code.multiline {
-    display: block;
-  }
-}
-.form-footer {
-  display: flex;
-  justify-content: flex-end;
-}
-
 kbd {
   background: #2c2c2c;
   border-radius: 4px;
